Move FIDO2 Cache to tenant space
